# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuoteSummary {
  id: number;
  quoteNumber: string;
  customerName: string;
  projectName: string;
  module: string;
  status: string;
  total: number;
  margin: number;
  updatedAt: string;
}

export interface DashboardSummary {
  totalQuotes: number;
  draftQuotes: number;
  readyQuotes: number;
  totalQuoted: number;
  averageMargin: number;
  recentQuotes: QuoteSummary[];
}

export interface EvChargerInputs {
  chargerQuantity: number;
  chargerOutputAmps: number;
  circuitAmps: string;
  chargerSupply: string;
  connection: string;
  routeLength: number;
  wiringMethod: string;
  location: string;
  panelManufacturer: string;
  panelSpace: string;
  breakerRequirement: string;
  access: string;
  permit: string;
  loadManagement: string;
  disconnect: string;
  surgeProtection: string;
  panelModifications: string;
  difficulty: string;
  notes: string;
}

export interface AssemblyLine {
  id: string;
  category: string;
  description: string;
  quantity: number;
  unit: string;
  unitCost: number;
  extendedCost: number;
  source: string;
}

export interface PricingInput {
  materialCost: number;
  laborCost: number;
  materialMarkup: number;
  calculatedSellingPrice: number;
  finalSellingPrice: number;
  /** @nullable */
  laborOverride: number | null;
  /** @nullable */
  sellingPriceOverride: number | null;
}

export type PricingSummary = PricingInput & {
  grossProfit: number;
  grossMargin: number;
};

export type Quote = QuoteSummary & ({
  /** @nullable */
  customerEmail: string | null;
  jobInputs: EvChargerInputs;
  assembly: AssemblyLine[];
  pricing: PricingSummary;
  proposalDescription: string;
  createdAt: string;
});

export interface QuoteInput {
  /** @minLength 1 */
  customerName: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @minLength 1 */
  projectName: string;
  module: string;
  jobInputs: EvChargerInputs;
  assembly?: AssemblyLine[];
  pricing?: PricingInput;
  /** @minLength 1 */
  proposalDescription: string;
}

export interface QuoteUpdate {
  customerName?: string;
  /** @nullable */
  customerEmail?: string | null;
  projectName?: string;
  status?: string;
  pricing?: PricingInput;
  proposalDescription?: string;
}

export interface PriceBookItem {
  id: number;
  category: string;
  item: string;
  unit: string;
  unitCost: number;
  isDefault: boolean;
  updatedAt: string;
}

export interface PriceBookItemUpdate {
  unitCost: number;
}

export interface CompanySettings {
  companyName: string;
  laborRate: number;
  materialMarkup: number;
  targetMargin: number;
  defaultTaxRate: number;
}

export interface CompanySettingsUpdate {
  companyName?: string;
  laborRate?: number;
  materialMarkup?: number;
  targetMargin?: number;
  defaultTaxRate?: number;
}

export type ListQuotesParams = {
status?: string;
};